options-strategy-platform/requirements.txt
# AWS CDK Core Dependencies
aws-cdk-lib>=2.80.0
constructs>=10.0.0

# AWS SDK for Python
boto3>=1.26.0
botocore>=1.29.0

# Options Trading Specific
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Market Data and Financial Calculations
yfinance>=0.2.0
requests>=2.28.0

# Date and Time Handling
python-dateutil>=2.8.2
pytz>=2023.3

# JSON and Data Validation
pydantic>=2.0.0

# Logging and Monitoring
structlog>=23.1.0

# WebSocket for Real-time Updates
websockets>=11.0.0

# Mathematical Operations for Greeks
sympy>=1.11.0